Refactoring towards more DI
